# Conflicts:
#	.env
#	composer.lock
#	config/packages/reset_password.yaml
#	package-lock.json
#	src/Controller/ResetPasswordController.php

+++ src/Controller/ProfilController.php
@@ -8,6 +8,7 @@ use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\String\Slugger\SluggerInterface;
 
@@ -28,11 +29,12 @@ class ProfilController extends AbstractController
     }
 
     #[Route('/modifier/{id}', name:"app_modifier")]
-    public function Modify(int $id, Request $request, EntityManagerInterface $entityManager,
-                           ParticipantRepository $participantRepository, SluggerInterface $slugger ): Response
+    public function modify(int $id, Request $request, EntityManagerInterface $entityManager,
+                           ParticipantRepository $participantRepository, SluggerInterface $slugger,
+                           UserPasswordHasherInterface $userPasswordHasher ): Response
     {
         $user = $this->getUser();
-        $participant = $participantRepository->findOneBy(['id' => $id]);
+        $participant = $participantRepository->find($id);
         $participantForm = $this->createForm(RegistrationFormType::class, $participant);
         $participantForm->handleRequest($request);
             if ($user!=null and $user->getUserIdentifier()==$participant->getUserIdentifier()){
@@ -46,12 +48,18 @@ class ProfilController extends AbstractController
                             $this->getParameter('image_directory'),
                             $newFilename);
                         if (file_exists('./uploads/image/'.$user->getImageFilename())) {
-                            if ($user->getImageFilename() != 'noimage.jpg'){
+                            if ($user->getImageFilename() != 'noimage.jpg' and $user->getImageFilename() !=""){
                                 unlink('./uploads/image/'.$user->getImageFilename());
                             }
                         }
                         $user->setImageFilename($newFilename);
                     }
+                    $user->setPassword(
+                        $userPasswordHasher->hashPassword(
+                            $user,
+                            $participantForm->get('plainPassword')->getData()
+                        )
+                    );
                     $entityManager->persist($participant);
                     $entityManager->flush();
 
@@ -61,7 +69,7 @@ class ProfilController extends AbstractController
             }
             else{
                 $this->addFlash('error', 'Vous ne pouvez pas accéder à cette page !');
-                return $this->redirectToRoute('app_home');
+                return $this->redirectToRoute('sortie_home');
             }
             return $this->render('registration/register.html.twig', [
                 'title' => 'Modifier le profil',
